+++ src/pages/profile/components/post/banner/BannerInfo.tsx
