+++ src/pages/profile/ui/ExpertDetailPage.tsx
@@ -2,8 +2,8 @@ import AppBar from '@shared/ui/AppBar.tsx';
 import { HeaderAction } from '@shared/types';
 import IcHamburger from '@icon/icon-hamburger.svg';
 import styled from 'styled-components';
-import ExpertTabs from '@my/components/ExpertTabs.tsx';
-import ExpertProfile from '@my/components/profile/ExpertProfile.tsx';
+import ExpertTabs from '@pages/profile/components/ExpertTabs.tsx';
+import ExpertProfile from '@pages/profile/components/profile/ExpertProfile.tsx';
 
 export const ExpertDetailPage = () => {
   const rightHeaderActions: HeaderAction[] = [
